+++ vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
             input: [
                 'resources/css/app.css',
                 'resources/js/app.ts',
+                'resources/js/filament/rich-text-editor.ts',
                 'resources/css/filament/admin/theme.css',
             ],
             refresh: true,